fix(zip): parse per-entry ZIP64 extra field (0x0001) in central dir (BUG-059)
强制 ZIP64 打包的词典 zip(即使<4GB)的中央目录条目用 0xFFFFFFFF 哨兵,
旧 parse_central_directory 一遇哨兵即 return false → zip.open 失败 →
导入报 unsupported format。按 APPNOTE 4.5.3 解析 0x0001 扩展字段取 64 位真值;
ZipEntry 尺寸字段拓宽为 uint64_t。真凶样本: （大修館）明鏡国語辞典［第二版］.zip

+++ native/hoshidicts/hoshidicts_src/zip/zip.hpp
@@ -10,8 +10,8 @@
 struct ZipEntry {
   std::string name;
   uint16_t compression_method;
-  uint32_t compressed_size;
-  uint32_t uncompressed_size;
+  uint64_t compressed_size;    // ZIP64 may store true size via 0x0001 extra
+  uint64_t uncompressed_size;  // ZIP64 may store true size via 0x0001 extra
   size_t data_offset;
 };
